+++ packages/adminui-pub/_src/app/ux/UBReportEditor.js
@@ -364,7 +364,7 @@ Ext.define('UB.ux.UBReportEditor', {
                 "advlist autolink lists charmap print preview hr anchor pagebreak", //link image
                 "searchreplace wordcount visualblocks visualchars code ", //fullscreen
                 "insertdatetime nonbreaking table contextmenu directionality", //save
-                "emoticons template textcolor colorpicker image paste"  //templateEditor paste
+                "emoticons template textcolor colorpicker image imagetools paste"  //templateEditor paste
             ],
             toolbar1: "undo redo | bold italic underline | alignleft aligncenter alignright alignjustify | formatselect fontsizeselect | pageOrientation | Colontitle | borderL borderR borderT borderB borderE borderA | forecolor | bullist numlist outdent indent |",
             contextmenu: 'link image inserttable | cell row column deletetable | rowTemplate',
@@ -400,9 +400,50 @@ Ext.define('UB.ux.UBReportEditor', {
              table  : {title : 'Table' , items : 'inserttable tableprops deletetable | cell row column'},
              tools  : {title : 'Tools' , items : 'spellchecker code'}
              },*/
-            toolbar_items_size: 'small'
+            toolbar_items_size: 'small',
             //height: '100%'
             //resize: true
+          automatic_uploads: false,
+          // here we add custom filepicker only to Image dialog
+            file_picker_types: 'image',
+          // and here's our custom image picker
+          file_picker_callback: function(cb, value, meta) {
+            var input = document.createElement('input');
+            input.setAttribute('type', 'file');
+            input.setAttribute('accept', 'image/*');
+
+            // Note: In modern browsers input[type="file"] is functional without
+            // even adding it to the DOM, but that might not be the case in some older
+            // or quirky browsers like IE, so you might want to add it to the DOM
+            // just in case, and visually hide it. And do not forget do remove it
+            // once you do not need it anymore.
+
+            input.onchange = function() {
+              var file = this.files[0];
+
+              // Note: Now we need to register the blob in TinyMCEs image blob
+              // registry. In the next release this part hopefully won't be
+              // necessary, as we are looking to handle it internally.
+              let reader = new FileReader();
+              let blob = new Blob([file], {type: file.type});
+              reader.addEventListener('loadend', function () {
+                var id = 'blobid' + (new Date()).getTime();
+                var blobCache = tinymce.activeEditor.editorUpload.blobCache;
+                // MPV - we need a base64 representation to embed image to document
+                let b64 = reader.result.split(',', 2)[1];
+                var blobInfo = blobCache.create(id, blob, b64);
+                blobCache.add(blobInfo);
+                // call the callback and populate the Title field with the file name
+                cb(blobInfo.blobUri(), { title: file.name });
+              })
+              // reader.addEventListener('error', function (event) {
+              //   reject(event)
+              // })
+              reader.readAsDataURL(blob)
+            }
+            input.click();
+          }
+
         }, me.tinyMCEEditorConfig || {} );
 
         me.tinyMCEConfig.reportEditor = me;