+++ packages/adminui-pub/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - UBBoxSelect now accept ubRequest as a config parameter (for store creation)
+- UBReportEditor can insert image from file (Insert -> Image -> click on button with photo)
+- UBReportEditor added build-in image editor (click on image to actiate)
 
 ### Fixed
 - since form definitian now evaluated only once (HMR) both `EntityGridPanel` & `ubdetailgrid` now accept `customActionas` as a Ext.Action config (not a class instance)